app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["irocshers <develop.iam@rocshers.com>"]
packages = [{include = "tt_web_app"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
python-decouple = "^3.7"
psycopg2 = "^2.9.5"
celery = "^5.2.7"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pytest = "^7.2.1"
isort = "^5.12.0"

[tool.black]
skip-string-normalization = true
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
